#17 Removed amq, still synchronous
