+++ build.gradle
@@ -9,8 +9,8 @@ allprojects  {
 
 subprojects {
   apply plugin: 'java'
-  sourceCompatibility = 1.7
-  targetCompatibility = 1.7
+  sourceCompatibility = 1.8
+  targetCompatibility = 1.8
 
   task packageSources(type: Jar) {
   classifier = 'sources'
