--- contract/pom.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-  
-    <parent>
-        <artifactId>parent</artifactId>
-        <groupId>com.navid.lazylogin</groupId>
-        <version>1.4-SNAPSHOT</version>
-    </parent>
-  
-    <groupId>com.navid.lazylogin</groupId>
-    <artifactId>contract</artifactId>
-    <name>LazyLogin-contract</name>
-  
-</project>
