--- domain/pom.xml
@@ -1,20 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    
-    <parent>
-        <groupId>com.navid.lazylogin</groupId>
-        <artifactId>parent</artifactId>
-        <version>1.4-SNAPSHOT</version>
-    </parent>
-    
-    <groupId>com.navid.lazylogin</groupId>
-    <artifactId>domain</artifactId>
-    <name>LazyLogin-domain</name>
-    <dependencies>
-        <dependency>
-            <groupId>com.google.guava</groupId>
-            <artifactId>guava</artifactId>
-        </dependency>
-    </dependencies>
-</project>
