--- eventconsumer-amq-email/build.gradle
@@ -1,8 +1,0 @@
-
-description = 'LazyLogin-eventconsumer-amq-email'
-dependencies {
-    compile group: 'org.apache.geronimo.specs', name: 'geronimo-jms_1.1_spec', version:'1.0.1'
-    compile group: 'org.springframework', name: 'spring-context-support', version:'4.1.4.RELEASE'
-    compile group: 'org.slf4j', name: 'slf4j-api', version:'1.7.8'
-    runtime group: 'javax.mail', name: 'mail', version:'1.4.7'
-}
