--- eventconsumer-amq-email/pom.xml
@@ -1,35 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-  
-    <parent>
-        <artifactId>parent</artifactId>
-        <groupId>com.navid.lazylogin</groupId>
-        <version>1.4-SNAPSHOT</version>
-    </parent>
-  
-    <groupId>com.navid.lazylogin</groupId>
-    <artifactId>eventconsumer-amq-email</artifactId>
-    <name>LazyLogin-eventconsumer-amq-email</name>
-    
-    <dependencies>
-        <dependency>
-            <groupId>org.apache.geronimo.specs</groupId>
-            <artifactId>geronimo-jms_1.1_spec</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.springframework</groupId>
-            <artifactId>spring-context-support</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.slf4j</groupId>
-            <artifactId>slf4j-api</artifactId>
-        </dependency>
-        <!-- runtime -->
-        <dependency>
-            <groupId>javax.mail</groupId>
-            <artifactId>mail</artifactId>
-            <scope>runtime</scope>
-        </dependency>
-    </dependencies>
-</project>
