--- eventproducer-amq/build.gradle
@@ -1,11 +1,0 @@
-
-description = 'LazyLogin-eventproducer-amq'
-dependencies {
-  compile project(':eventproducer-api')
-  compile project(':domain')
-    compile group: 'org.springframework', name: 'spring-jms', version:'4.1.4.RELEASE'
-    compile group: 'org.springframework', name: 'spring-context', version:'4.1.4.RELEASE'
-    compile group: 'org.apache.geronimo.specs', name: 'geronimo-jms_1.1_spec', version:'1.0.1'
-    runtime group: 'commons-pool', name: 'commons-pool', version:'1.6'
-    runtime group: 'org.apache.activemq', name: 'activemq-pool', version:'5.10.0'
-}
