--- eventproducer-amq/src/main/resources/conf/config-amq.xml
@@ -1,23 +1,0 @@
-<beans
-    xmlns="http://www.springframework.org/schema/beans"
-    xmlns:amq="http://activemq.apache.org/schema/core"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xmlns:jms="http://www.springframework.org/schema/jms"
-    xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-2.0.xsd
-  http://activemq.apache.org/schema/core http://activemq.apache.org/schema/core/activemq-core.xsd
-http://www.springframework.org/schema/jms http://www.springframework.org/schema/jms/spring-jms-3.0.xsd">
-
-    <bean id="pooledJmsConnectionFactory" class="org.apache.activemq.pool.PooledConnectionFactory" destroy-method="stop">
-        <property name="connectionFactory"  >
-            <bean id="jmsSslConnectionFactory" class="org.apache.activemq.ActiveMQSslConnectionFactory">
-                <property name="brokerURL" value="${amq.url}" />
-                <property name="keyStore" value="${security.keystore.loc}" />
-                <property name="keyStorePassword" value="${security.keystore.password}" />
-                <property name="trustStore" value="${security.truststore.loc}" />
-                <property name="trustStorePassword" value="${security.truststore.password}" />
-            </bean>
-        </property>
-    </bean>
-
- 
-</beans>
