+++ eventproducer-nojms/src/main/java/com/navid/lazylogin/eventproducer/NoJMSConfiguration.java
@@ -1,0 +1,14 @@
+package com.navid.lazylogin.eventproducer;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.ImportResource;
+
+/**
+ * Created by alberto on 8/15/15.
+ */
+@Configuration
+@ImportResource("classpath:conf/config-nojms-email.xml")
+public class NoJMSConfiguration {
+
+
+}
