+++ eventproducer-nojms/src/main/resources/conf/config-nojms-email.xml
@@ -6,11 +6,6 @@
     xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-2.0.xsd
   http://activemq.apache.org/schema/core http://activemq.apache.org/schema/core/activemq-core.xsd
 http://www.springframework.org/schema/jms http://www.springframework.org/schema/jms/spring-jms-3.0.xsd">
- 
-    <bean id="jmsTemplate" class="org.springframework.jms.core.JmsTemplate">
-        <property name="connectionFactory" ref="pooledJmsConnectionFactory" />
-        <property name="defaultDestination" ref="destination"/>
-    </bean>
     
     <bean id="mailSender" class="org.springframework.mail.javamail.JavaMailSenderImpl">
         <property name="host" value="${email.host}" />
@@ -31,16 +26,5 @@ http://www.springframework.org/schema/jms http://www.springframework.org/schema/
         <property name="from" value="${email.from}"/>
         <property name="subject" value="${email.subject}"/>
     </bean>
-    
-    <bean id="emailNotifier" class="com.navid.lazylogin.jmsconsumer.EmailNotifier"/>
-    
-    <bean id="destination" class="org.apache.activemq.command.ActiveMQQueue">
-        <constructor-arg value="jmsExample" />
-    </bean>
-     
-    <!-- The Spring message listener container configuration -->
-    <jms:listener-container container-type="default" connection-factory="pooledJmsConnectionFactory" acknowledge="auto">
-        <jms:listener destination="jmsExample" ref="emailNotifier" method="onMessage" />
-    </jms:listener-container>
 
 </beans>
