--- logic/pom.xml
@@ -1,84 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    
-    <parent>
-        <groupId>com.navid.lazylogin</groupId>
-        <artifactId>parent</artifactId>
-        <version>1.4-SNAPSHOT</version>
-    </parent>
-    
-    <groupId>com.navid.lazylogin</groupId>
-    <artifactId>logic</artifactId>
-    <name>LazyLogin-logic</name>
-    
-    <dependencies>
-        
-        <!-- Internal dependencies -->
-        <dependency>
-            <groupId>com.navid.lazylogin</groupId>
-            <artifactId>domain</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>com.navid.lazylogin</groupId>
-            <artifactId>eventproducer-api</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>com.navid.lazylogin</groupId>
-            <artifactId>persistence-api</artifactId>
-        </dependency>
-        
-        <!-- External dependencies -->
-        <dependency>
-            <groupId>org.jdto</groupId>
-            <artifactId>jdto</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.slf4j</groupId>
-            <artifactId>slf4j-api</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.cxf</groupId>
-            <artifactId>cxf-rt-rs-service-description</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.springframework</groupId>
-            <artifactId>spring-tx</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.springframework</groupId>
-            <artifactId>spring-beans</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>javax.ws.rs</groupId>
-            <artifactId>javax.ws.rs-api</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>javax.annotation</groupId>
-            <artifactId>javax.annotation-api</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.springframework</groupId>
-            <artifactId>spring-context</artifactId>
-        </dependency>
-        
-        <!-- Test dependencies -->
-        <dependency>
-            <groupId>org.springframework</groupId>
-            <artifactId>spring-test</artifactId>
-            <scope>test</scope>
-        </dependency>
-        
-        <dependency>
-            <groupId>org.slf4j</groupId>
-            <artifactId>jcl-over-slf4j</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.springframework</groupId>
-            <artifactId>spring-web</artifactId>
-        </dependency>
-    </dependencies>
-
-    
-
-</project>
