+++ springboot/build.gradle
@@ -3,13 +3,18 @@ buildscript {
         mavenCentral()
     }
     dependencies {
-        //classpath("org.springframework.boot:spring-boot-gradle-plugin:1.2.5.RELEASE")
+        classpath("org.springframework.boot:spring-boot-gradle-plugin:1.2.5.RELEASE")
         classpath('se.transmode.gradle:gradle-docker:1.2')
     }
 }
 
-apply plugin: 'war'
+apply plugin: 'spring-boot'
 apply from: 'docker.gradle'
+
+jar {
+    baseName = 'gamemanager'
+    version =  '0.1.0'
+}
 
 description = 'LazyLogin-springboot-endpoint'
 dependencies {
@@ -49,8 +54,7 @@ dependencies {
     compile group: 'javax.servlet', name: 'jstl', version: '1.2'
     compile group: 'org.hsqldb', name: 'hsqldb', version: '2.3.2'
     runtime project(':persistence-jpa')
-    runtime project(':eventproducer-amq')
-    runtime project(':eventconsumer-amq-email')
+    runtime project(':eventproducer-nojms')
     runtime group: 'com.navid.spring', name: 'addon-jms-performance', version: '1.3'
     runtime(group: 'org.springframework.data', name: 'spring-data-commons', version: '1.9.1.RELEASE') {
         exclude(module: 'commons-logging')
